requirements.txt
lxml
openpyxl
pyyaml
pyqt6
